sdks/python/arc-llamaindex/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arc-llamaindex"
version = "0.1.0"
description = "LlamaIndex integration for the ARC protocol -- capability-scoped FunctionTool and QueryEngineTool with vector-collection scoping"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Backbay Industries" },
]
keywords = ["arc", "llamaindex", "agent", "tool", "rag", "security", "capability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    "arc-sdk-python>=0.1.0",
    "llama-index-core>=0.11,<1",
    "pydantic>=2.5,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "httpx>=0.27,<1",
    "respx>=0.21,<1",
    "mypy>=1.10",
    "ruff>=0.5",
]

[tool.uv.sources]
arc-sdk-python = { path = "../arc-sdk-python" }

[tool.hatch.build.targets.wheel]
packages = ["src/arc_llamaindex"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_return_any = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]
